start working image translator

+++ main.py
@@ -47,16 +47,20 @@ def check_login(message):
 
 @bot.message_handler(content_types=['photo'])
 def contact(message):
-    fileID = message.photo[-1].file_id
-    file_info = bot.get_file(fileID)
-    downloaded_file = bot.download_file(file_info.file_path)
-    with open(f"{fileID}.jpg", 'wb') as new_file:
-        new_file.write(downloaded_file)
-    img = Image.open(f'{fileID}.jpg')
-    result = pytesseract.image_to_string(img)
-    translation = Translator().translate(result, dest="fa")
-    os.system(f"rm -rf {fileID}.jpg")
-    bot.send_message(message.chat.id, translation.text)
+    user_id = message.from_user.id
+    if DataBaseManagerUser.check_login(user_id):
+        fileID = message.photo[-1].file_id
+        file_info = bot.get_file(fileID)
+        downloaded_file = bot.download_file(file_info.file_path)
+        with open(f"{fileID}.jpg", 'wb') as new_file:
+            new_file.write(downloaded_file)
+        img = Image.open(f'{fileID}.jpg')
+        result = pytesseract.image_to_string(img)
+        translation = Translator().translate(result, dest="fa")
+        os.system(f"rm -rf {fileID}.jpg")
+        bot.send_message(message.chat.id, translation.text)
+    else:
+        bot.send_message(message.chat.id, 'Please Use /register for login in this bot')
 
 
 bot.polling()
